backend/pyproject.toml
[project]
name = "papersearch-agent-backend"
version = "0.1.0"
description = "Multi-agent scientific paper search backend with FastAPI and MCP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "sqlalchemy>=2.0.36",
  "aiosqlite>=0.20.0",
  "pydantic-settings>=2.6.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt==3.2.2",
  "email-validator>=2.2.0",
  "python-multipart>=0.0.12",
  "langgraph>=0.2.39",
  "langchain-openai>=0.2.6",
  "langchain-core>=0.3.15",
  "mcp>=1.2.0",
  "orjson>=3.10.10",
  "httpx>=0.27.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.1",
  "mypy>=1.13.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
